Add models.py
The news model contains only the title, picture, date and time of creation, the author. The content of the news contains in the blocks of news. Each block can contain one or all of the filled-in fields. Depending on how full it is, it changes the visual presentation in the template. For example, the text is wrapped around the pictures if it exists or picture is displayed in full screen if the text is missing. In turn, the video is always displayed in full screen and etc.