Add service.ts
The universal code to work with API servers. The ServiceBase is base class implemented to work with APIs and handle responses from any server. Children classes look like PartnerService, where we set interface for data, method, and URL for any request. This approach significantly reduces the time of development every next service for the API.